metrohash uses a 64 bit seed, not a 32 bit seed

+++ metrohash.h
@@ -32,20 +32,20 @@
 // TODO: msvc compat
 
 // MetroHash 64-bit hash functions
-void metrohash64_1(const uint8_t * key, uint64_t len, uint32_t seed, uint8_t * out);
-void metrohash64_2(const uint8_t * key, uint64_t len, uint32_t seed, uint8_t * out);
+void metrohash64_1(const uint8_t * key, uint64_t len, uint64_t seed, uint8_t * out);
+void metrohash64_2(const uint8_t * key, uint64_t len, uint64_t seed, uint8_t * out);
 
 // MetroHash 64-bit hash functions using HW CRC instruction
-void metrohash64crc_1(const uint8_t * key, uint64_t len, uint32_t seed, uint8_t * out);
-void metrohash64crc_2(const uint8_t * key, uint64_t len, uint32_t seed, uint8_t * out);
+void metrohash64crc_1(const uint8_t * key, uint64_t len, uint64_t seed, uint8_t * out);
+void metrohash64crc_2(const uint8_t * key, uint64_t len, uint64_t seed, uint8_t * out);
 
 // MetroHash 128-bit hash functions
-void metrohash128_1(const uint8_t * key, uint64_t len, uint32_t seed, uint8_t * out);
-void metrohash128_2(const uint8_t * key, uint64_t len, uint32_t seed, uint8_t * out);
+void metrohash128_1(const uint8_t * key, uint64_t len, uint64_t seed, uint8_t * out);
+void metrohash128_2(const uint8_t * key, uint64_t len, uint64_t seed, uint8_t * out);
 
 // MetroHash 128-bit hash functions using HW CRC instruction
-void metrohash128crc_1(const uint8_t * key, uint64_t len, uint32_t seed, uint8_t * out);
-void metrohash128crc_2(const uint8_t * key, uint64_t len, uint32_t seed, uint8_t * out);
+void metrohash128crc_1(const uint8_t * key, uint64_t len, uint64_t seed, uint8_t * out);
+void metrohash128crc_2(const uint8_t * key, uint64_t len, uint64_t seed, uint8_t * out);
 
 
 /* rotate right idiom recognized by compiler*/
